move light and materials into sub-folder
