+++ src/index.js
@@ -15,8 +15,8 @@ import GLSL from './gl/glsl';
 import ShaderProgram from './gl/shader_program';
 import VertexData from './gl/vertex_data';
 import Texture from './gl/texture';
-import Material from './material';
-import Light from './light';
+import Material from './lights/material';
+import Light from './lights/light';
 import WorkerBroker from './utils/worker_broker';
 import Task from './utils/task';
 import {StyleManager} from './styles/style_manager';
