+++ src/lights/ambient_light.glsl
