+++ src/styles/style.js
@@ -5,8 +5,8 @@ import FeatureSelection from '../selection';
 import ShaderProgram from '../gl/shader_program';
 import VBOMesh from '../gl/vbo_mesh';
 import Texture from '../gl/texture';
-import Material from '../material';
-import Light from '../light';
+import Material from '../lights/material';
+import Light from '../lights/light';
 import {RasterTileSource} from '../sources/raster';
 import log from '../utils/log';
 import mergeObjects from '../utils/merge';
